src/requirements.txt
PyGithub==1.59.1
python-dotenv==1.0.0